Move `MockKeyVault` to hasher
