+++ .github/workflows/main.yml
@@ -60,7 +60,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -e pixl_core
-          pip install -e pytest-pixl
           pip install hasher/[test]
 
       - name: Run tests
